+++ worker/src/twitter.rs
@@ -29,21 +29,6 @@ pub struct TwitterResponse {
     pub includes: TwitterResponseIncludes,
 }
 
-pub fn extract_signature(tweet: String) -> Result<(String, String)> {
-    let mut sig_target = "".to_string();
-    for line in tweet.split('\n').collect::<Vec<&str>>() {
-        if line.starts_with("sig:") {
-            if sig_target != "" {
-                return Ok((sig_target, line[4..].to_string().clone()));
-            } else {
-                return Err(anyhow!("Signature target is empty."));
-            }
-        }
-        sig_target = format!("{}{}\n", sig_target, line);
-    }
-    Err(anyhow!("Signature not found in tweet."))
-}
-
 pub async fn retrieve_tweet(api_token: String, tweet_id: String) -> Result<TwitterResponse> {
     let mut headers = HeaderMap::new();
     headers.insert(AUTHORIZATION, format!("Bearer {}", &api_token).parse()?);
